+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 === Head
 
+* Add namespace option to BEM preset pattern.
 * Improve strictness of SUIT preset pattern: enforce proper camelCasing.
 
 === 1.2.0 (October 12, 2015)